requirements.txt
numpy==1.13.3
mlxtend==0.17.2
pyfim==6.28
tabulate==0.8.7
pandas==1.0.3
mock==4.0.2
